testing model on data stored through Spark

+++ Scripts/preprocessing.py
@@ -4,11 +4,13 @@ from pyspark.sql import SparkSession
 from pyspark.sql.types import *
 from pyspark.sql.functions import isnan, when, count, col, to_date, udf, lower, regexp_replace, transform
 from pyspark.sql.types import DateType
-from pyspark.ml.feature import StopWordsRemover, Tokenizer, CountVectorizer, VectorAssembler, StringIndexer,IndexToString
+from pyspark.ml.feature import StopWordsRemover, Tokenizer, CountVectorizer, VectorAssembler, StringIndexer, \
+    IndexToString
 from pyspark.ml import Pipeline, PipelineModel
 from pyspark.ml.classification import LogisticRegression
 from pyspark.ml.evaluation import BinaryClassificationEvaluator, MulticlassClassificationEvaluator
 from pyspark.mllib.evaluation import MulticlassMetrics
+
 
 class ModelTraining():
 
@@ -17,36 +19,49 @@ class ModelTraining():
             .appName("SparkByExamples.com") \
             .getOrCreate()
         self.csv_PATH = "/home/saranoor/Data/spark_project/twitter_sentiment_analysis/Data_Model_Training/training_n_validation/twitter_training.csv"
-        self.model_path= '/home/saranoor/Data/spark_project/twitter_sentiment_analysis/Models/'
+        self.model_path = '/home/saranoor/Data/spark_project/twitter_sentiment_analysis/Models/'
 
     def read_data(self, PATH, type):
-         if type=='train':
-             self.df=self.spark.read. \
+        if type == 'train':
+            self.df = self.spark.read. \
                 option('headers', True). \
                 option('inferSchema', True). \
                 csv(PATH)
-         elif type=='test':
-             schema = StructType(
-                 [
-                  StructField('id', IntegerType(), True),
-                  StructField('dontknow', StringType(), True),
-                  StructField('str_target', StringType(), True),
-                  StructField('text', StringType(), True),
-                  ]
-             )
-             self.df_test=self.spark.read. \
+        elif type == 'test':
+            schema = StructType(
+                [
+                    StructField('id', IntegerType(), True),
+                    StructField('dontknow', StringType(), True),
+                    StructField('str_target', StringType(), True),
+                    StructField('text', StringType(), True),
+                ]
+            )
+            self.df_test = self.spark.read. \
                 option('headers', True). \
                 schema(schema). \
                 csv(PATH)
 
+        elif type == 'spark':
+            schema = StructType(
+            [
+                StructField('value', StringType(), True),
+                StructField('id', StringType(), True),
+                StructField('text', StringType(), True),
+            ]
+            )
+            self.df_spark = self.spark.read. \
+                option("header", 'true'). \
+                option("inferSchema", 'true').\
+                csv('/home/saranoor/Data/spark_project/twitter_sentiment_analysis/output/filesink_4/spark_data.csv')
+
     def cleaning(self, type):
-        if type=='train':
+        if type == 'train':
             # data = self.df.selectExpr("_c0 as target","id","date","flag","user","text")
             self.df = self.df.toDF("id", "dontknow", "str_target", "text")
             ### finding out outliers
             ### count of null and none values
             self.df.select([count(when(isnan(c) | col(c).isNull(), True)).alias(c) for c in self.df.columns]).show()
-            self.df=self.df.na.drop()
+            self.df = self.df.na.drop()
             ### replace null values/missing values
             ###convert string to datetime
             # func = udf(lambda x: datetime.strptime(x, "%a %b %d %H:%M:%S PDT %Y"), DateType())
@@ -57,13 +72,15 @@ class ModelTraining():
             self.df = self.df.withColumn("text", regexp_replace(col("text"), r'[^a-z0-9]', ' '))
             # removing extra spaces
             self.df = self.df.withColumn("text", regexp_replace(col("text"), r"\s\s+", ' '))
-        else:
+
+        elif type == 'test':
             # data = self.df.selectExpr("_c0 as target","id","date","flag","user","text")
-            self.df_test= self.df_test.toDF("id", "dontknow", "str_target", "text")
+            self.df_test = self.df_test.toDF("id", "dontknow", "str_target", "text")
             ### finding out outliers
             ### count of null and none values
-            self.df_test.select([count(when(isnan(c) | col(c).isNull(), True)).alias(c) for c in self.df_test.columns]).show()
-            self.df_test=self.df_test.na.drop()
+            self.df_test.select(
+                [count(when(isnan(c) | col(c).isNull(), True)).alias(c) for c in self.df_test.columns]).show()
+            self.df_test = self.df_test.na.drop()
             ### replace null values/missing values
             ###convert string to datetime
             # func = udf(lambda x: datetime.strptime(x, "%a %b %d %H:%M:%S PDT %Y"), DateType())
@@ -75,16 +92,23 @@ class ModelTraining():
             # removing extra spaces
             self.df_test = self.df_test.withColumn("text", regexp_replace(col("text"), r"\s\s+", ' '))
 
+        elif type == 'spark':
+            self.df_spark = self.df_spark.na.drop()
+            self.df_spark = self.df_spark.withColumn("text", lower(col("text")))
+            self.df_spark = self.df_spark.withColumn("text", regexp_replace(col("text"), r'[^a-z0-9]', ' '))
+            self.df_spark = self.df_spark.withColumn("text", regexp_replace(col("text"), r"\s\s+", ' '))
+            self.df_spark.printSchema()
+            self.df_spark.show()
     def Pipeline(self):
         # convert target to int value
         stringIndexer = StringIndexer(inputCol="str_target", outputCol="target").fit(self.df)
 
-        #tokenzing the data
+        # tokenzing the data
         tokenizer = Tokenizer(inputCol="text", outputCol="text_tokenized")
-        #self.df=tokenizer.transform(self.df)
+        # self.df=tokenizer.transform(self.df)
 
         # removing stopwords
-        remover=StopWordsRemover(inputCol="text_tokenized",outputCol="text_clean")
+        remover = StopWordsRemover(inputCol="text_tokenized", outputCol="text_clean")
         # self.df=remover.transform(self.df)
         # self.df.show()
         # countvectorizer
@@ -93,49 +117,56 @@ class ModelTraining():
         # cv_model=cv.fit(self.df)
         # self.df=cv_model.transform(self.df)
 
-        #vectorAssembler
-        va=VectorAssembler(inputCols=['features'],
-                                  outputCol='attributes')
+        # vectorAssembler
+        va = VectorAssembler(inputCols=['features'],
+                             outputCol='attributes')
 
-        #training model
-        model=LogisticRegression(featuresCol='attributes', labelCol='target')
+        # training model
+        model = LogisticRegression(featuresCol='attributes', labelCol='target')
 
-        #indextoString
+        # indextoString
         labelConverter = IndexToString(inputCol="prediction", outputCol="predictedLabel",
-                                       labels=stringIndexer.labels)#creating pipeline
+                                       labels=stringIndexer.labels)  # creating pipeline
 
-        pipeline=Pipeline().setStages([stringIndexer, tokenizer, remover, cv, va, model, labelConverter])
-        pipeline_model=pipeline.fit(self.df)
-        self.df_transform=pipeline_model.transform(self.df)
+        pipeline = Pipeline().setStages([stringIndexer, tokenizer, remover, cv, va, model, labelConverter])
+        pipeline_model = pipeline.fit(self.df)
+        self.df_transform = pipeline_model.transform(self.df)
 
         self.df_transform.select('attributes', 'target', 'prediction').show()
-        #self.df_transform.show()
+        # self.df_transform.show()
 
-        lr_model=pipeline_model.stages[5]
-        trainingSummary=lr_model.summary
+        lr_model = pipeline_model.stages[5]
+        trainingSummary = lr_model.summary
         print("Accuracy of model is:" + str(trainingSummary.accuracy))
 
-        evaluator = MulticlassClassificationEvaluator(labelCol="target",predictionCol="prediction")
+        evaluator = MulticlassClassificationEvaluator(labelCol="target", predictionCol="prediction")
         print("Train Accuracy " + str(evaluator.evaluate(self.df_transform, {evaluator.metricName: "accuracy"})))
 
-        predictionAndLabels = self.df_transform.select("prediction","target").rdd
+        predictionAndLabels = self.df_transform.select("prediction", "target").rdd
         multi_metrics = MulticlassMetrics(predictionAndLabels)
         precision_score = multi_metrics.weightedPrecision
         recall_score = multi_metrics.weightedRecall
-        print('Precision and Recall are:',precision_score, recall_score)
+        print('Precision and Recall are:', precision_score, recall_score)
 
-        pipeline_model.write().overwrite().save(self.model_path+'pyspark-log-reg-model')
+        pipeline_model.write().overwrite().save(self.model_path + 'pyspark-log-reg-model')
 
     def evaluation(self):
-        self.read_data("/home/saranoor/Data/spark_project/twitter_sentiment_analysis/Data_Model_Training/training_n_validation/twitter_validation.csv", 'test')
-        load_model = PipelineModel.load(self.model_path+'pyspark-log-reg-model')
-        self.cleaning('test')
-        self.df_test_transform = load_model.transform(self.df_test)
-        self.df_test_transform.select('prediction','target').show()
-        evaluator = MulticlassClassificationEvaluator(labelCol="target",predictionCol="prediction")
-        print("Test Accuracy " + str(evaluator.evaluate(self.df_test_transform, {evaluator.metricName: "accuracy"})))
-        predictionAndLabels = self.df_test_transform.select("prediction","target").rdd
-        multi_metrics = MulticlassMetrics(predictionAndLabels)
-        precision_score = multi_metrics.weightedPrecision
-        recall_score = multi_metrics.weightedRecall
-        print('Precision and Recall are:',precision_score, recall_score)
+        self.read_data(
+            "/home/saranoor/Data/spark_project/twitter_sentiment_analysis/Data_Model_Training/training_n_validation/twitter_validation.csv",
+            'test')
+        load_model = PipelineModel.load(self.model_path + 'pyspark-log-reg-model')
+        # self.cleaning('test')
+        # self.df_test_transform = load_model.transform(self.df_test)
+        # self.df_test_transform.select('prediction','target').show()
+        # evaluator = MulticlassClassificationEvaluator(labelCol="target",predictionCol="prediction")
+        # print("Test Accuracy " + str(evaluator.evaluate(self.df_test_transform, {evaluator.metricName: "accuracy"})))
+        # predictionAndLabels = self.df_test_transform.select("prediction","target").rdd
+        # multi_metrics = MulticlassMetrics(predictionAndLabels)
+        # precision_score = multi_metrics.weightedPrecision
+        # recall_score = multi_metrics.weightedRecall
+        # print('Precision and Recall are:',precision_score, recall_score)
+
+        # testing on twitter data stored through SPARK
+        self.cleaning('spark')
+        self.df_spark_transform = load_model.transform(self.df_spark)
+        self.df_spark_transform.select('prediction').show()
